% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allplot.R
\name{allplot}
\alias{allplot}
\title{Plot all}
\usage{
allplot(res0, target = "hs", biomass.scale = 1000,
  pngfile = "../buri.png", detail.plot = 1:3)
}
\description{
単位はcatch at ageの尾数が100万尾、waaがgの場合、重量の単位がちょうどトンになるようになっている。
}
