% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-msy.R
\name{est.MSY}
\alias{est.MSY}
\title{Estimate MSY}
\usage{
est.MSY(vpares, farg, seed = 1, n.imputation = 1, nyear = 50,
  eyear = 0, FUN = mean, N = 1000, onlylower.pgy = FALSE,
  is.small = FALSE, is.Kobe = FALSE, is.5perlower = FALSE,
  optim.method = "optimize", max.target = "catch.mean",
  calc.yieldcurve = TRUE, Blimit = 0, trace.multi = c(seq(from = 0,
  to = 0.9, by = 0.1), 1, seq(from = 1.1, to = 2, by = 0.1), 3:5, 7, 20,
  100), is.plot = TRUE, PGY = NULL, B0percent = NULL)
}
\description{
Estimate MSY
}
