% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurervpa-package.R
\docType{package}
\name{futurervpa-package}
\alias{futurervpa}
\alias{futurervpa-package}
\title{futurervpa: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
\keyword{internal}
