% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kobemat.R
\name{get.kobemat}
\alias{get.kobemat}
\title{kobe.matrixの計算}
\usage{
get.kobemat(fout, N = fout$input$N, nyear = fout$input$nyear,
  Btarget = 0, fmulti = seq(from = 0.3, to = 1, by = 0.1))
}
\description{
Pr(B<Btarget)のみを返す単純なやつ
}
