% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-perform.R
\name{get.perform}
\alias{get.perform}
\title{Get perform}
\usage{
get.perform(fout0, Blimit = 0, longyear = 50, smallcatch = 0.5,
  N = NULL, shortyear = c(3, 5, 10), tmp.year = NULL)
}
\description{
Get perform
}
