% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-future.R
\name{plot.futures}
\alias{plot.futures}
\alias{plot.future}
\title{Plot future}
\usage{
\method{plot}{futures}(fres.list, conf = c(0.1, 0.5, 0.9),
  target = "SSB", legend.text = "", xlim.tmp = NULL, y.scale = 1)

\method{plot}{future}(fres0, ylim.tmp = NULL, xlim.tmp = NULL,
  vpares = NULL, what = c(TRUE, TRUE, TRUE), conf = 0.1,
  N.line = 0, label = c("Biomass", "SSB", "Catch"), is.legend = TRUE,
  add = FALSE, col = NULL, ...)
}
\description{
Plot future
}
