% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-kobe.R
\name{plot.kobe}
\alias{plot.kobe}
\alias{plot.kobemat}
\title{Plot kobe}
\usage{
\method{plot}{kobe}(vpares, Bref, Uref, plot.history = FALSE,
  is.plot = FALSE, pickU = "", pickB = "", ylab.tmp = "U/Umsy",
  xlab.tmp = "SSB/SSBmsy", title.tmp = "")

\method{plot}{kobemat}(xx, title.name = "", line = 0)
}
\description{
Plot kobe
}
