% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-vpa.R
\name{read.vpa}
\alias{read.vpa}
\title{結果の入力}
\usage{
read.vpa(tfile, caa.label = "catch at age",
  maa.label = "maturity at age", waa.label = "weight at age",
  waa.biomass.label = "weight at age for biomass calculation",
  waa.catch.label = "weight at age for catch calculation",
  M.label = "M at age", faa.label = "fishing mortality at age",
  Fc.label = "Current F", naa.label = "numbers at age",
  Blimit = NULL, Pope = TRUE, fc.year = NULL)
}
\description{
結果の入力
}
