% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr-data.R
\name{get.SRdata}
\alias{get.SRdata}
\alias{plot.SRdata}
\title{SR data}
\usage{
get.SRdata(vpares, R.dat = NULL, SSB.dat = NULL,
  years = as.numeric(colnames(vpares$naa)))

\method{plot}{SRdata}(SRdata)
}
\description{
SR data
}
