% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-msy2.R
\name{est.MSY2}
\alias{est.MSY2}
\title{Estimate MSY}
\usage{
est.MSY2(vpares, N = 1000, res1 = NULL, sim0 = NULL, nyear = NULL,
  pgy = 0.9, lim = 0.6, ban = 0.1, mY = 5, long.term = 20,
  Fmsy.max = 3, Fmsy.step = 0.1, thin = 1, inc = 1,
  SRtype = "L2", fm = 5, tol = NULL, AutoCor = FALSE,
  AutoCorOut = FALSE, current.resid = 0,
  future.function.name = "future.vpa1", seed = 1)
}
\description{
Estimate MSY
}
