% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-vpa.R
\name{future.vpa}
\alias{future.vpa}
\title{将来予測関数}
\usage{
future.vpa(res0, currentF = NULL, multi = 1, nyear = 10,
  Pope = res0$input$Pope, outtype = "FULL", multi.year = 1,
  start.year = NULL, ABC.year = NULL, waa.year = NULL,
  maa.year = NULL, M.year = NULL, seed = NULL, strategy = "F",
  HCR = NULL, beta = NULL, delta = NULL, Blim = 0, Bban = 0,
  plus.group = res0$input$plus.group, N = 1000, silent = FALSE,
  is.plot = TRUE, random.select = NULL, pre.catch = NULL,
  rec.new = NULL, recfunc = HS.recAR, rec.arg = list(upper.ssb = Inf,
  upper.recruit = Inf), Frec = NULL, waa = NULL, waa.catch = NULL,
  maa = NULL, M = NULL, waa.multi = "opt", waa.optyear = 2011:2013,
  replace.rec.year = 2012, F.sigma = 0, waa.fun = FALSE,
  naa0 = NULL, eaa0 = NULL, ssb0 = NULL, faa0 = NULL,
  add.year = 0, det.run = TRUE)
}
\description{
multiのオプションは管理後のFのmultiplier（管理前後でselectivityが同じ）
}
