% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurervpa-package.R
\docType{package}
\name{futurervpa-package}
\alias{futurervpa}
\alias{futurervpa-package}
\title{futurervpa: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/futurervpa}
  \item Report bugs at \url{https://github.com/heavywatal/futurervpa/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
\keyword{internal}
